requirements.txt
alembic
appdirs
asgiref
black
certifi
click
colorama
distlib
fastapi
FastAPI-SQLAlchemy
filelock
greenlet
h11
Mako
MarkupSafe
mypy-extensions
pathspec
pipenv
psycopg2
pydantic
python-dateutil
python-dotenv
python-editor
regex
six
SQLAlchemy
starlette
toml
typing-extensions
uvicorn
virtualenv
virtualenv-clone




